Revert "adding build support for bundles with bicep steps (#127)"
This reverts commit a8ca8f7d9cdab8e718b45c07c34ce42cdfd7df5d.
