+++ pkg/commands/build_bundle.go
@@ -3,7 +3,6 @@ package commands
 import (
 	"fmt"
 
-	"github.com/massdriver-cloud/mass/pkg/bicep"
 	"github.com/massdriver-cloud/mass/pkg/bundle"
 	"github.com/massdriver-cloud/mass/pkg/restclient"
 	"github.com/massdriver-cloud/mass/pkg/terraform"
@@ -32,11 +31,6 @@ func BuildBundle(buildPath string, b *bundle.Bundle, c *restclient.MassdriverCli
 			}
 		case "helm":
 			continue
-		case "bicep":
-			err = bicep.GenerateFiles(buildPath, step.Path, b, fs)
-			if err != nil {
-				return err
-			}
 		default:
 			return fmt.Errorf("%s is not a supported provisioner", step.Provisioner)
 		}
