--- pkg/mockfilesystem/testdata/bicep/massdriver.yaml
@@ -1,63 +1,0 @@
-schema: draft-07
-name: draft-node
-description: A resource that can be used to visually design architecture without provisioning real infrastructure.
-ref: github.com/massdriver-cloud/massdriver-bundles
-access: public
-type: bundle
-
-steps:
-  - path: src
-    provisioner: bicep
-
-params:
-  examples:
-    - __name: Network
-      resource_type: Network
-  required:
-    - resource_type
-  properties:
-    resource_name:
-      type: string
-      title: Resource Name
-      description: An immutable name field
-      $md.immutable: true
-    resource_type:
-      type: string
-      title: Resource Type
-      description: The type of resource
-    foo:
-      type: object
-      title: Foo
-      description: A map of Foos
-      required:
-        - bar
-      properties:
-        bar:
-          type: integer
-          title: A whole number
-          description: Testing numbers
-          default: 1
-        qux:
-          type: integer
-          title: A whole number that is not required
-          description: Testing numbers
-          minimum: 2
-
-connections:
-  required:
-    - draft_node_foo
-  properties:
-    draft_node_foo:
-      $ref: massdriver/draft-node
-
-artifacts:
-  required:
-    - draft_node
-  properties:
-    draft_node:
-      $ref: massdriver/draft-node
-
-ui:
-  ui:order:
-    - resource_name
-    - "*"
