+++ pkg/terraform/generate_files.go
@@ -23,11 +23,6 @@ var mdVars = map[string]interface{}{
 	},
 }
 
-const (
-	ParamsFile = "_params.auto.tfvars.json"
-	ConnsFile  = "_connections.auto.tfvars.json"
-)
-
 func GenerateFiles(buildPath, stepPath string, b *bundle.Bundle, fs afero.Fs) error {
 	err := generateTfVarsFiles(buildPath, stepPath, b, fs)
 
@@ -35,7 +30,7 @@ func GenerateFiles(buildPath, stepPath string, b *bundle.Bundle, fs afero.Fs) er
 		return err
 	}
 
-	devParamPath := path.Join(buildPath, stepPath, ParamsFile)
+	devParamPath := path.Join(buildPath, stepPath, bundle.ParamsFile)
 
 	err = transpileAndWriteDevParams(devParamPath, b, fs)
 
@@ -43,7 +38,7 @@ func GenerateFiles(buildPath, stepPath string, b *bundle.Bundle, fs afero.Fs) er
 		return fmt.Errorf("error compiling dev params: %w", err)
 	}
 
-	err = transpileConnectionVarFile(path.Join(buildPath, stepPath, ConnsFile), b, fs)
+	err = transpileConnectionVarFile(path.Join(buildPath, stepPath, bundle.ConnsFile), b, fs)
 
 	if err != nil {
 		return err
