+++ pkg/terraform/transpile_connection_var_files.go
@@ -2,21 +2,40 @@ package terraform
 
 import (
 	"encoding/json"
+	"fmt"
 
 	"github.com/massdriver-cloud/mass/pkg/bundle"
-	"github.com/massdriver-cloud/mass/pkg/inputvars"
 	"github.com/spf13/afero"
 )
 
 func transpileConnectionVarFile(path string, b *bundle.Bundle, fs afero.Fs) error {
-	values, err := inputvars.TranspileConnectionVarFile(path, b, fs)
+	emptyConnections := checkEmptySchema(b.Connections)
+
+	if emptyConnections {
+		err := afero.WriteFile(fs, path, []byte("{}"), 0755)
+
+		if err != nil {
+			return err
+		}
+
+		return nil
+	}
+
+	existingConnectionsVars, err := getExistingVars(path, fs)
 
 	if err != nil {
 		return err
 	}
 
-	bytes, err := json.MarshalIndent(values, "", "    ")
+	connectionsSchemaProperties, ok := b.Connections["properties"].(map[string]interface{})
 
+	if !ok {
+		return fmt.Errorf("expected connections schema properties to be an object")
+	}
+
+	values, _ := setValuesIfNotExists(connectionsSchemaProperties, existingConnectionsVars, nil)
+
+	bytes, err := json.MarshalIndent(values, "", "    ")
 	if err != nil {
 		return err
 	}
